Address message chaining in moveToNextMachine
Add a short method in the Machine class
named addJob which puts a job in the Machine's
job queue. This fixes a minor message chain in
moveToNextMachine. Also created a new method
in SimulationSpecification named getJobTasks
which returns the number of tasks a Job is specified
to have. This alleviates a message chain in the setupJobs
method. Addresses #7