Fix message chains in setUpJobs
Add two methods in SimulationSpecification
for retrieving the task time for a task
in a job, and the machine for a task in a job.
These methods are now both called in setUpJobs
instead of using message chains to achieve the
same functionality. In addition, a Machine method
addJob replaces another message chain in the same
method. Addresses #7