Create and use MachineShopSimulator get and set methods
Get and Set methods for the instance variables
in MachineShopSimulator have been created and are
now being used in Machine and RunMachineShop instead
of directly accessing the variables. As a result,
the instance variables have been changed back to private.
Message chains have also been addressed in RunMachineShop
by using these methods and retrieving the variables on seperate
lines. Addresses #9 #10